extracting docker images update process from apt service update
